Fixed up unit tests and package requirements.
