+++ .gitignore
@@ -1,10 +1,6 @@
-*build
-yarn-error.log
 .DS_Store
+*/.DS_Store
 .sass-cache
 node_modules
-bower_components
-css
-stylesheets
-scripts
-src/js/latex-parser.js
+*build
+coverage
